BinanceSpot Development

+++ src/ExchangeApi/Result/Spot/SpotOrderBookEntry.php
@@ -18,7 +18,7 @@ class SpotOrderBookEntry
         return $this->price;
     }
 
-    public function getQuantity(): int
+    public function getQuantity(): float
     {
         return $this->quantity;
     }
